pyproject.toml
[project]
name = "fontdownloader"
version = "0.1.0"
description = "Download fonts from Google Fonts, get the difference between two font lists, and more"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    'requests',
    'fonttools',
    'python-dotenv'
]
